refactor: fix Infrastructure DIP gap, move AssetValueRequest to DTOs
Audited Financial.Infrastructure's own interfaces and their consumers.
Found two real issues:

- StatusInvestFinanceService implemented IFinanceService but was wired
  and consumed by its concrete type (BondAssetPriceFetcher took a
  StatusInvestFinanceService parameter directly), unlike its sibling
  GoogleFinanceService which correctly goes through the interface.
  BondAssetPriceFetcher now depends on IFinanceService; DI still wires
  the specific StatusInvestFinanceService singleton via a factory
  registration (same pattern already used for IRepository in this file),
  since IFinanceService intentionally has two non-interchangeable
  implementations (Google for standard/crypto, StatusInvest for bonds)
  rather than a resolve-any-one collection like IAssetPriceFetcher.

- AssetValueRequest was a plain DTO sitting inside Interfaces/, not a
  contract. Moved to Financial.Infrastructure/DTOs/AssetValueRequestDTO.cs,
  renamed to match the DTO naming convention already used throughout
  Financial.Application.
